Use VolumeInfo to help restore the PV.
Add VolumeInfo for left PVs during backup.
