+++ changelogs/unreleased/7138-blackpiglet
@@ -1,0 +1,1 @@
+Use VolumeInfo to help restore the PV.
