+++ pkg/backup/backup_test.go
@@ -71,8 +71,9 @@ func TestBackedUpItemsMatchesTarballContents(t *testing.T) {
 	req := &Request{
 		Backup:           defaultBackup().Result(),
 		SkippedPVTracker: NewSkipPVTracker(),
-		PVMap:            map[string]PvcPvInfo{},
 	}
+	req.VolumesInformation.InitPVMap()
+
 	backupFile := bytes.NewBuffer([]byte{})
 
 	apiResources := []*test.APIResource{
