+++ pkg/backup/item_backupper.go
@@ -699,28 +699,18 @@ func (ib *itemBackupper) addVolumeInfo(obj runtime.Unstructured, log logrus.Fiel
 		return err
 	}
 
-	if ib.backupRequest.PVMap == nil {
-		ib.backupRequest.PVMap = make(map[string]PvcPvInfo)
-	}
-
 	pvcName := ""
 	pvcNamespace := ""
 	if pv.Spec.ClaimRef != nil {
 		pvcName = pv.Spec.ClaimRef.Name
 		pvcNamespace = pv.Spec.ClaimRef.Namespace
-
-		ib.backupRequest.PVMap[pvcNamespace+"/"+pvcName] = PvcPvInfo{
-			PVCName:      pvcName,
-			PVCNamespace: pvcNamespace,
-			PV:           *pv,
-		}
 	}
 
-	ib.backupRequest.PVMap[pv.Name] = PvcPvInfo{
+	ib.backupRequest.VolumesInformation.InsertPVMap(pv.Name, PvcPvInfo{
 		PVCName:      pvcName,
 		PVCNamespace: pvcNamespace,
 		PV:           *pv,
-	}
+	})
 	return nil
 }
