+++ pkg/controller/restore_controller.go
@@ -56,6 +56,7 @@ import (
 	kubeutil "github.com/vmware-tanzu/velero/pkg/util/kube"
 	"github.com/vmware-tanzu/velero/pkg/util/logging"
 	"github.com/vmware-tanzu/velero/pkg/util/results"
+	"github.com/vmware-tanzu/velero/pkg/volume"
 )
 
 // nonRestorableResources is an exclusion list  for the restoration process. Any resources
@@ -520,6 +521,16 @@ func (r *restoreReconciler) runValidatedRestore(restore *api.Restore, info backu
 		return errors.Wrap(err, "fail to fetch CSI VolumeSnapshots metadata")
 	}
 
+	backupVolumeInfoMap := make(map[string]volume.VolumeInfo)
+	volumeInfos, err := backupStore.GetBackupVolumeInfos(restore.Spec.BackupName)
+	if err != nil || volumeInfos == nil {
+		restoreLog.Infof("Backup %s doesn't have volumeinfos metadata file.", restore.Spec.BackupName)
+	} else {
+		for _, volumeInfo := range volumeInfos.VolumeInfos {
+			backupVolumeInfoMap[volumeInfo.PVName] = volumeInfo
+		}
+	}
+
 	restoreLog.Info("starting restore")
 
 	var podVolumeBackups []*api.PodVolumeBackup
@@ -537,6 +548,7 @@ func (r *restoreReconciler) runValidatedRestore(restore *api.Restore, info backu
 		ResourceModifiers:    resourceModifiers,
 		DisableInformerCache: r.disableInformerCache,
 		CSIVolumeSnapshots:   csiVolumeSnapshots,
+		VolumeInfoMap:        backupVolumeInfoMap,
 	}
 	restoreWarnings, restoreErrors := r.restorer.RestoreWithResolvers(restoreReq, actionsResolver, pluginManager)
